requirements.txt
django
Pillow==5.2.0
openpyxl
django-crispy-forms
